Changelog addition, took out the ReportABugMailView into its own file
