+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use only relays added in RelayView for sending and receiving events
 
 - Add PostHog analytics
+- Open an email compose view controller for support@planetary.social 
 
 ## [0.1 (6)] - 2023-03-08Z
